FIx :  Order edit setup -4

+++ app/src/main/java/com/example/groceryshoppingapp/models/Item.kt
@@ -1,21 +1,3 @@
-//package com.example.groceryshoppingapp.models
-//
-//import android.os.Parcelable
-//import kotlinx.parcelize.Parcelize
-//
-//@Parcelize
-//data class Item(
-//    val id: String,
-//    val name: String,
-//    val description: String,
-//    val price: Double,
-//    val stockQuantity: Int,
-//    val createdAt: String? = null,
-//    val createdBy: String? = null,
-//    val shopid: String,
-//    val imageUrl: String? = null
-//
-//) : Parcelable
 package com.example.groceryshoppingapp.models
 
 import android.os.Parcelable
@@ -29,7 +11,7 @@ data class Item(
     val description: String? = null,
     val price: Double,
     @SerializedName("quantity")
-    val stockQuantity: Int,
+    val stockQuantity: Double,
     val createdAt: String? = null,
     val createdBy: String? = null,
     @SerializedName("shopId")
